requirements.txt
django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
environs==14.1.1
psycopg2-binary==2.9.10
pdfkit==1.0.0
qrcode==8.0